【modify】 主键修改整形Integer

+++ zyx-dubbo-live/src/main/java/com/zyx/rpc/live/LiveInfoFacade.java
@@ -35,7 +35,7 @@ public interface LiveInfoFacade {
 	 * @author MrDeng
 	 * @update 2016年6月22日 上午10:30:44
 	 */
-	public LiveInfo getById(Long id);
+	public LiveInfo getById(Integer id);
 	/**
 	 * 
 	 * @param liveInfoVo
@@ -66,7 +66,7 @@ public interface LiveInfoFacade {
 	 * @author MrDeng
 	 * @update 2016年6月22日 上午10:30:06
 	 */
-	public void delete(Long id);
+	public void delete(Integer id);
 	
 	/**
 	 * 
@@ -77,7 +77,7 @@ public interface LiveInfoFacade {
 	 * @author MrDeng
 	 * @update 2016年6月27日 下午6:18:28
 	 */
-	public String getLiveUrl(Long liveId);
+	public String getLiveUrl(Integer liveId);
 	
 	/**
 	 * 获取首推
